change landing-page files structure
