+++ packages/client/src/components/LandingPage/Header/Header.tsx
@@ -1,3 +1,5 @@
+import './Header.css';
+
 export default function () {
   return (
     <header>
