+++ packages/client/src/components/LandingPage/LandingPage.css
@@ -1,9 +1,4 @@
-@import './styles/helpers.css';
 @import './styles/fonts.css';
-@import './styles/components/Header.css';
-@import './styles/components/Main.css';
-@import './styles/components/Footer.css';
-@import './styles/components/Wallpaper.css';
 
 .landing-page-container {
   --color-cold-white: #e7f6f6;
@@ -24,7 +19,6 @@
 
   width: 100%;
   height: 100%;
-
   font-size: var(--font-size-normal);
   color: var(--font-color-default);
 }
