--- packages/client/src/components/LandingPage/styles/helpers.css
@@ -1,11 +1,0 @@
-@define-placeholder flex-column-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-@define-placeholder flex-row-container {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
